* move Scene from util to its own package for further grouping related scene-management classes together * make addGameObject non-static, because there will be multiple superclass unrelated functions soon
